[core] move metrics to plugin (#577)
