+++ src/main/java/io/javalin/core/plugin/PluginLifecyles.kt
@@ -4,10 +4,10 @@ import io.javalin.Javalin
 
 /** Extend [Plugin] with a new lifecycle */
 interface PluginLifecycleInit {
-  /**
-   * Initialize properties and event listener.
-   * This will be called before any handler is registered.
-   * It is not allowed to register handler during this lifecycle.
-   */
-  fun init(app: Javalin)
+    /**
+     * Initialize properties and event listener.
+     * This will be called before any handler is registered.
+     * It is not allowed to register handler during this lifecycle.
+     */
+    fun init(app: Javalin)
 }
