--- src/main/java/io/javalin/plugin/metrics/MetricsProvider.kt
@@ -1,9 +1,0 @@
-/*
- * Javalin - https://javalin.io
- * Copyright 2017 David Åse
- * Licensed under Apache 2.0: https://github.com/tipsy/javalin/blob/master/LICENSE
- */
-
-package io.javalin.plugin.metrics
-
-enum class MetricsProvider { NONE, MICROMETER }
